requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
google==3.0.0
googlesearch-python==1.2.3
idna==3.4
requests==2.28.2
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
tzdata==2023.3
urllib3==1.26.15
